Moved setters out into abstract class
