+++ pangocairohelpers/text_path/__init__.py
@@ -1,3 +1,4 @@
 from .text_path_glyph_item import TextPathGlyphItem  # noqa
+from .text_path_abstract import TextPathAbstract  # noqa
 from .text_path import TextPath  # noqa
 from .upright_text_path import UprightTextPath  # noqa
